Added pixel art generation

+++ Structs/Position.go
@@ -2,7 +2,7 @@ package Structs
 
 // Position describes position of a tile. Simply a three-dimensional vector
 type Position struct {
-	X float32 `json:"x"`
-	Y float32 `json:"y"`
-	Z float32 `json:"z"`
+	X int `json:"x"`
+	Y int `json:"y"`
+	Z int `json:"z"`
 }
